@albertodiazdorado 2025-01-20
Hi @Airfooox , do you think you could have a look into this?