@Airfooox 2025-03-13
:tada: This PR is included in version 1.5.0 :tada:

The release is available on:
- [npm package (@latest dist-tag)](https://www.npmjs.com/package/cron-validate/v/1.5.0)
- [GitHub release](https://github.com/Airfooox/cron-validate/releases/tag/v1.5.0)

Your **[semantic-release](https://github.com/semantic-release/semantic-release)** bot :package::rocket: